refactor(getVersion): move to device-core
