+++ apps/ledger-live-desktop/src/renderer/modals/UpdateFirmwareModal/helpers/createFirmwareUpdateSteps.ts
@@ -1,4 +1,4 @@
-import { isDeviceLocalizationSupported } from "@ledgerhq/live-common/manager/localization";
+import { isDeviceLocalizationSupported } from "@ledgerhq/live-common/device-core/commands/use-cases/isDeviceLocalizationSupported";
 import { FirmwareUpdateContext } from "@ledgerhq/types-live";
 import { DeviceModelId } from "@ledgerhq/devices";
 import StepResetDevice, { StepResetFooter } from "../steps/00-step-reset-device";
