+++ libs/ledger-live-common/src/device-core/commands/use-cases/isDeviceLocalizationSupported.test.ts
@@ -1,5 +1,5 @@
 import { DeviceModelId } from "@ledgerhq/devices";
-import { isDeviceLocalizationSupported } from "./localization";
+import { isDeviceLocalizationSupported } from "./isDeviceLocalizationSupported";
 const { nanoS, nanoSP, nanoX, stax } = DeviceModelId;
 
 test("isDeviceLocalizationSupported", () => {
