+++ libs/ledger-live-common/src/device-core/commands/use-cases/isDeviceLocalizationSupported.ts
