+++ libs/ledger-live-common/src/device/use-cases/getVersionUseCase.ts
@@ -2,4 +2,4 @@ export {
   getVersion,
   isBootloaderVersionSupported,
   isHardwareVersionSupported,
-} from "../../hw/getVersion";
+} from "../../device-core/commands/use-cases/getVersion";
